Add entry generator function
Created a function that generates new entries on my page, based on an array of the projects. So now, instead of copy-pasting and changing the code on the index.html page, I can simply add the name of the project in the array and the entry will show on my progress page!